+++ src/main/java/com/hubo/gillajabi/crawl/domain/service/AbstarctCrawlService.java
@@ -1,0 +1,9 @@
+package com.hubo.gillajabi.crawl.domain.service;
+
+import java.util.List;
+
+public abstract class AbstarctCrawlService {
+    public abstract List<?> crawlCourse();
+
+    public abstract List<?> crawlTheme();
+}
