+++ src/main/java/com/hubo/gillajabi/crawl/infrastructure/config/RoadProperties.java
@@ -1,10 +1,8 @@
 package com.hubo.gillajabi.crawl.infrastructure.config;
 
-import com.hubo.gillajabi.crawl.domain.constant.ApiProperties;
 import com.hubo.gillajabi.crawl.domain.constant.CityName;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
-import org.hibernate.validator.constraints.Length;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
